requirements.txt
# Core
requests>=2.32
python-dotenv>=1.0

# Email agent (Step 4)
pdfplumber>=0.11

# Dashboard (Step 5)
flask>=3.0

# Scheduling (Step 8)
apscheduler>=3.10

# Scraping (Step 9)
playwright>=1.44
